requirements.txt
nni
torchattacks
torchtoolbox
mne
mne_connectivity
torchmetrics
tqdm
numpy
pandas
scipy
scikit-learn
lmdb
einops
torch_geometric
xmltodict
networkx
PyWavelets
spectrum
moabb
pytorch-lightning